[Code-Refactoring]: Code Organize, Bug Fixed and Reuseable Function Implemented
