+++ index.js
@@ -1,6 +1,6 @@
 // Import File
 const app = require("./app");
-const connectDB = require("./config/DB");
+const connectDB = require("./src/config/DB");
 
 // listen server port
 const PORT = process.env.PORT || 5000;
